Hepatitis project
The 'hcvdat0' dataset from kaggle was used to train clustering algorithms (Kmeans and Kmeans++) to predict whether someone has no illness, hepatitis, cirrhosis or fibrosis. Initially, exploratory data analysis was performed, in the form of aggregate statistics and visualisations, to gain insights about the data. Principal component analysis (PCA) was then carried out, using the elbow method, to reduce the dimensionality of the data while preserving as much of its information as possible. The Kmeans and Kmeans++ models were then trained (with and without the PCA data) and the number of clusters optimised appropriately, using inertia values and the elbow method. The models were then evaluated based on their inertias and the best model was chosen.